test: remove DeleteFactHandlerTests file
